Our new egcs is fuzzy (and nows the types of printf!)

+++ platform/emulator/dp_table.cc
@@ -269,8 +269,8 @@ void OwnerTable::print(){
         next = oe->uOB.oExt;
         printf("<%d>\t %s\t", i, toC(PO_getValue(oe)));
         while(next != NULL){
-          printf("ex:%d %d#%d ", ctr, next->getCredit(0),
-                 next->getCredit(1));
+          printf("ex:%d %d#%d ", (int) ctr, (int) next->getCredit(0),
+                 (int) next->getCredit(1));
           ctr ++;
           next = next->getNext();}
         printf("\n");}
@@ -279,7 +279,7 @@ void OwnerTable::print(){
           printf("<%d>\t %s\t PERSISTENT\n", i, toC(PO_getValue(oe)));
         else
           printf("<%d>\t %s\t %d\n", i, toC(PO_getValue(oe)),
-                 oe->uOB.credit);}
+                 (int) oe->uOB.credit);}
 #ifdef XXDEBUG_PERDIO
       getOwner(i)->print();
     } else{
@@ -1032,44 +1032,44 @@ void BorrowEntry::print_entry(int nr) {
            toC(PO_getValue(this)), netaddr.index);
     return;
   case PO_EXTENDED|PO_SLAVE|PO_MASTER|PO_BIGCREDIT:
-    sprintf(primCred, "%d(slave)", getSlave()->getMaster()->primCredit);
+    sprintf(primCred, "%d(slave)", (int) getSlave()->getMaster()->primCredit);
     next = getSlave()->getMaster()->uSOB.oce;
     *secCred = '\0';
     temp = secCred;
     while(next != NULL){
       temp += strlen(temp);
-      sprintf(temp, "ex:%d %d#%d ", ctr, next->getCredit(0),
-              next->getCredit(1));
+      sprintf(temp, "ex:%d %d#%d ", ctr, (int) next->getCredit(0),
+              (int) next->getCredit(1));
       ctr ++;
       next = next->getNext();}
     break;
   case PO_EXTENDED|PO_SLAVE|PO_MASTER:
-    sprintf(primCred, "%d(slave)", getSlave()->getMaster()->primCredit);
-    sprintf(secCred, "%d", getSlave()->getMaster()->uSOB.secCredit);
+    sprintf(primCred, "%d(slave)", (int) getSlave()->getMaster()->primCredit);
+    sprintf(secCred, "%d", (int) getSlave()->getMaster()->uSOB.secCredit);
     break;
   case PO_EXTENDED|PO_SLAVE:
-    sprintf(primCred, "%d", getSlave()->primCredit);
-    sprintf(secCred, "%d", getSlave()->uSOB.secCredit);
+    sprintf(primCred, "%d", (int) getSlave()->primCredit);
+    sprintf(secCred, "%d", (int) getSlave()->uSOB.secCredit);
     break;
   case PO_EXTENDED|PO_MASTER|PO_BIGCREDIT:
-    sprintf(primCred, "%d", getMaster()->primCredit);
+    sprintf(primCred, "%d", (int) getMaster()->primCredit);
     next = getMaster()->uSOB.oce;
     *secCred = '\0';
     temp = secCred;
     while(next != NULL){
       temp += strlen(temp);
-      sprintf(temp, "ex:%d %d#%d ", ctr, next->getCredit(0),
-              next->getCredit(1));
+      sprintf(temp, "ex:%d %d#%d ", ctr, (int) next->getCredit(0),
+              (int) next->getCredit(1));
       ctr ++;
       next = next->getNext();}
     break;
   case PO_EXTENDED|PO_MASTER:
-    sprintf(primCred, "%d", getMaster()->primCredit);
-    sprintf(secCred, "%d", getMaster()->uSOB.secCredit);
+    sprintf(primCred, "%d", (int) getMaster()->primCredit);
+    sprintf(secCred, "%d", (int) getMaster()->uSOB.secCredit);
     break;
   case PO_NONE:
     secCred = "0";
-    sprintf(primCred, "%d", uOB.credit);
+    sprintf(primCred, "%d", (int) uOB.credit);
     break;
   default:
     Assert(0);}
